+++ src/Integration.php
@@ -22,7 +22,7 @@ use WP_User;
  * Company: Pronamic
  *
  * @author  Remco Tolsma
- * @version 2.0.0
+ * @version 2.0.8
  * @since   1.0.0
  */
 class Integration extends AbstractIntegration {
@@ -67,7 +67,7 @@ class Integration extends AbstractIntegration {
 		$function = array( $this, 'next_payment_delivery_date' );
 
 		if ( ! \has_filter( 'pronamic_pay_subscription_next_payment_delivery_date', $function ) ) {
-			add_filter( 'pronamic_pay_subscription_next_payment_delivery_date', $function, 10, 2 );
+			\add_filter( 'pronamic_pay_subscription_next_payment_delivery_date', $function, 10, 2 );
 		}
 
 		add_filter( 'pronamic_payment_provider_url_mollie', array( $this, 'payment_provider_url' ), 10, 2 );